Fixed clanGL to build on OS X

+++ Sources/GL/Platform/OSX/input_device_provider_osxkeyboard.h
@@ -103,7 +103,7 @@ public:
 	}
 
 	/// \brief Sets the position of the device.
-	void set_position(int x, int y) { return; }
+	void set_position(float x, float y) { return; }
 
 /// \}
 /// \name Implementation
